make version a Swath attribute
